# ButtonComponent remove EntityReference & optimize add/remove speed.