--- a/src/scene/graph-node.js
+++ b/src/scene/graph-node.js
@@ -1232,8 +1232,9 @@ class GraphNode extends EventHandler {
             this.localRotation.setFromMat4(invParentWtm).mul(rotation);
         }
 
-        if (!this._dirtyLocal)
+        if (!this._dirtyLocal) {
             this._dirtifyLocal();
+        }
     }
 
     /**