@mvaligursky 2025-01-06
cherry picked to main_v2